codec: fix regression from msgpack container len limit fix which failed compile on 32-bit machines
Fixes #433

+++ codec/msgpack.go
@@ -293,7 +293,7 @@ func (e *msgpackEncDriver[T]) writeContainerLen(ct msgpackContainerType, l int) 
 	} else if l < 65536 {
 		e.w.writen1(ct.b16)
 		e.w.writen2(bigen.PutUint16(uint16(l)))
-	} else if l <= mpMaxLen {
+	} else if uint(l) <= mpMaxLen {
 		e.w.writen1(ct.b32)
 		e.w.writen4(bigen.PutUint32(uint32(l)))
 	} else {
